# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fabiq.SmartFactory.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Oee/work-order/{workOrderId}": {
      "get": {
        "tags": [
          "Oee"
        ],
        "parameters": [
          {
            "name": "workOrderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "plannedProductionMinutes",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 120
            }
          },
          {
            "name": "idealCycleTimeSeconds",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OeeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OeeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OeeDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Traceability/part/{partId}": {
      "get": {
        "tags": [
          "Traceability"
        ],
        "parameters": [
          {
            "name": "partId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PartTraceabilityDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartTraceabilityDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartTraceabilityDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Traceability/work-order/{workOrderId}": {
      "get": {
        "tags": [
          "Traceability"
        ],
        "parameters": [
          {
            "name": "workOrderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderTraceabilityDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderTraceabilityDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderTraceabilityDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/WorkOrders/{workOrderId}": {
      "get": {
        "tags": [
          "WorkOrders"
        ],
        "parameters": [
          {
            "name": "workOrderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OeeDto": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "nullable": true
          },
          "plannedProductionMinutes": {
            "type": "number",
            "format": "double"
          },
          "runningMinutes": {
            "type": "number",
            "format": "double"
          },
          "idealCycleTimeSeconds": {
            "type": "number",
            "format": "double"
          },
          "goodCount": {
            "type": "integer",
            "format": "int32"
          },
          "scrapCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "availability": {
            "type": "number",
            "format": "double"
          },
          "performance": {
            "type": "number",
            "format": "double"
          },
          "quality": {
            "type": "number",
            "format": "double"
          },
          "oee": {
            "type": "number",
            "format": "double"
          },
          "calculatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PartTraceabilityDto": {
        "type": "object",
        "properties": {
          "partId": {
            "type": "string",
            "nullable": true
          },
          "finalStatus": {
            "type": "string",
            "nullable": true
          },
          "currentMachineId": {
            "type": "string",
            "nullable": true
          },
          "currentLineId": {
            "type": "string",
            "nullable": true
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "route": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartTraceabilityStepDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PartTraceabilityStepDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "nullable": true
          },
          "machineId": {
            "type": "string",
            "nullable": true
          },
          "lineId": {
            "type": "string",
            "nullable": true
          },
          "workOrderId": {
            "type": "string",
            "nullable": true
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "defectType": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "WorkOrderDto": {
        "type": "object",
        "properties": {
          "workOrderId": {
            "type": "string",
            "nullable": true
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "lineId": {
            "type": "string",
            "nullable": true
          },
          "plannedQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "goodCount": {
            "type": "integer",
            "format": "int32"
          },
          "scrapCount": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkOrderPartSummaryDto": {
        "type": "object",
        "properties": {
          "partId": {
            "type": "string",
            "nullable": true
          },
          "finalStatus": {
            "type": "string",
            "nullable": true
          },
          "goodQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "scrapQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "lastMachineId": {
            "type": "string",
            "nullable": true
          },
          "lastEventType": {
            "type": "string",
            "nullable": true
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "WorkOrderTraceabilityDto": {
        "type": "object",
        "properties": {
          "workOrderId": {
            "type": "string",
            "nullable": true
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "lineId": {
            "type": "string",
            "nullable": true
          },
          "plannedQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "goodCount": {
            "type": "integer",
            "format": "int32"
          },
          "scrapCount": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "partCount": {
            "type": "integer",
            "format": "int32"
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkOrderPartSummaryDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}